M1365 Separate new attribute into new component
