--- src/components/generic/InputAutocomplete/index.js
@@ -1,3 +1,0 @@
-import InputAutocomplete from './InputAutocomplete'
-
-export default InputAutocomplete
