+++ src/components/pages/ImageClassification/ImageAnnotationModal/ImageAnnotationModal.styles.tsx
@@ -171,7 +171,7 @@ export const PopupWrapperForRadio = styled.div`
 export const NewAttributeModalContentContainer = styled.div`
   display: flex;
   gap: ${theme.spacing.large};
-  min-height: 250px;
+  // min-height: 250px;
 `
 
 export const NewAttributeModalFooterContainer = styled.div`
