+++ src/components/pages/ImageClassification/ImageAnnotationModal/ImageAnnotationPopup/SelectAttributeFromClassifierGuesses.jsx
@@ -11,15 +11,12 @@ import { databaseSwitchboardPropTypes } from '../../../../../App/mermaidData/dat
 import { IconPlus } from '../../../../icons'
 import {
   LabelThatLooksLikeATh,
-  NewAttributeModalContentContainer,
   NewAttributeModalFooterContainer,
-  NewAttributeModalLabel,
   RowThatLooksLikeAnEvenTr,
 } from '../ImageAnnotationModal.styles'
 import { Select } from '../../../../generic/form'
 import { useSelectNewAttribute } from '../../useSelectNewAttribute'
-import InputAutocomplete from '../../../../generic/InputAutocomplete'
-import language from '../../../../../language'
+import NewAttributeModal from './NewAttributeModal'
 import Modal from '../../../../generic/Modal/Modal'
 
 const isClassified = ({ annotations }) => annotations.length > 0
@@ -157,6 +154,7 @@ const SelectAttributeFromClassifierGuesses = ({
         </Select>
       </RowThatLooksLikeAnEvenTr>
       {createPortal(
+        //modal will otherwise populate within the map container
         <Modal
           title="Select New Attribute"
           isOpen={
@@ -169,37 +167,13 @@ const SelectAttributeFromClassifierGuesses = ({
           maxWidth="fit-content"
           contentOverflowIsVisible
           mainContent={
-            <NewAttributeModalContentContainer>
-              <NewAttributeModalLabel htmlFor="benthic-attribute-autocomplete">
-                Benthic Attribute
-                <InputAutocomplete
-                  id="benthic-attribute-autocomplete"
-                  // eslint-disable-next-line jsx-a11y/no-autofocus
-                  autoFocus // IMPORTANT we should reconsider autofocus use. See: https://trello.com/c/4pe1zgS9/1331-accessibility-linting-issues-deferred
-                  aria-labelledby="benthic-attribute-label"
-                  options={benthicAttributeSelectOptions}
-                  onChange={({ value }) => setSelectedBenthicAttr(value)}
-                  value={selectedBenthicAttr}
-                  noResultsText={language.autocomplete.noResultsDefault}
-                />
-              </NewAttributeModalLabel>
-
-              <NewAttributeModalLabel htmlFor="growth-forms">
-                <span>Growth forms</span>
-                <Select
-                  id="growth-forms"
-                  label="Growth forms"
-                  onChange={(e) => setSelectedGrowthForm(e.target.value)}
-                >
-                  <option value=""></option>
-                  {growthFormSelectOptions.map((growthForm) => (
-                    <option key={growthForm.id} value={growthForm.id}>
-                      {growthForm.name}
-                    </option>
-                  ))}
-                </Select>
-              </NewAttributeModalLabel>
-            </NewAttributeModalContentContainer>
+            <NewAttributeModal
+              benthicAttributeSelectOptions={benthicAttributeSelectOptions}
+              setSelectedBenthicAttr={setSelectedBenthicAttr}
+              setSelectedGrowthForm={setSelectedGrowthForm}
+              selectedBenthicAttr={selectedBenthicAttr}
+              growthFormSelectOptions={growthFormSelectOptions}
+            />
           }
           footerContent={
             <NewAttributeModalFooterContainer>
